# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Architecture Knowledge Platform API
  version: 0.2.0
  x-akp-write-idempotency: >
    Mutating POST operations accept Idempotency-Key and MCP write tools require
    it. POST /v1/auth/session is intentionally exempt because it creates a
    Set-Cookie response that cannot be safely replayed.
  description: >
    HTTP contract for the routes registered by the current Fastify runtime.
    All /v1 operations require a scoped bearer API token or an unexpired
    HttpOnly web session and enforce the permission recorded in
    x-akp-permission. Session-authenticated writes also require X-CSRF-Token.
servers:
  - url: http://127.0.0.1:8080
    description: Local development default
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      x-akp-permission: knowledge:read
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/projects/scan:
    post:
      tags: [Projects]
      operationId: scanProject
      x-akp-permission: knowledge:propose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, rootPath]
              properties:
                slug: { type: string, minLength: 1 }
                rootPath: { type: string, minLength: 1 }
                spaceId: { type: string, format: uuid }
                commit: { type: string, pattern: "^[a-fA-F0-9]{40}$" }
                changedSince: { type: string, minLength: 1 }
      responses:
        "201": { $ref: "#/components/responses/ObjectResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/schema/dry-run:
    post:
      tags: [Governance]
      operationId: dryRunKnowledgeSchema
      x-akp-permission: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [candidateVersion]
              properties:
                spaceId: { type: string, format: uuid }
                candidateVersion: { type: string, minLength: 1 }
                requiredFrontmatterFields:
                  { type: array, items: { type: string } }
                allowedTypes: { type: array, items: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
      description: >
        Provisioned out of band with an explicit TokenScopeSet. Token scopes
        are intersected with the user's current memberships on every request.
    cookieAuth:
      type: apiKey
      in: cookie
      name: akp_session
      description: HttpOnly session carrying the effective TokenScopeSet snapshot.
...
  responses:
    ObjectResponse:
      description: Runtime response object
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, expired, revoked or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Permission, space or path scope denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code: { type: string }
      additionalProperties: true
...